Implemented routes write in new routing via SCR agent
